Loop idiom recognition for polynomial division.
The polynomial multiplication instructions can be used to speed the
calculation of remainder of polynomial division using Barrett's formula.
In this patch:

(1) Ported the recognition part of polynomial division from
HexagonLoopIdiomRecognition.

(2) Introduce the `riscv.pdivstepsr` intrinsic. If we find polynomial
division pattern, we emit it.

(3) If all conditions are met to apply Barrett's formula we expand
riscv.pdivstepsr intrinsic using `clmul` and `clmulh` intrinsics.
Otherwise, we emit an (unrolled) loop which what `riscv.pdivstepsr`
computed. This expansion happens in RISCVCodeGenPrepare.