Add test for clamp like case where the first select is a ([s/z]ext (icmp))
This ([s/z]ext (icmp)) is equivilant to (select icmp, [1/-1], 0), so can be
canonicalized too. This is generally not beneficial because after the canonicalization
we lose the ability replace one of the selects with ([s/z]ext (icmp)). However, it is
beneficial for this particular case:
```
  %old_cmp1 = icmp sgt i32 %x, C2
  %old_replacement = sext i1 %old_cmp1 to i32
  %old_cmp0 = icmp ult i32 %x, C0
  %r = select i1 %old_cmp0, i32 %x, i32 %old_replacement
it can be rewriten as more canonical pattern:
  %new_cmp2 = icmp sge i32 %x, C0
  %new_clamped_low = smax i32 %target_low, i32 %x
  %r = select i1 %new_cmp2, i32 -1, i32 %new_clamped_low
Iff 0 s<= C2 s<= C0
```
The select can be lowered to:
```
  %sext_cmp2 = sext i1 %new_cmp2 to i32
  %r = or i32 %sext_cmp2, i32 %new_clamped_low
```